fix(shared): clear the 18 pre-existing mypy errors in src/shared/python
main's quality-gate mypy step only type-checks files changed in a push, so
these 18 errors sat unnoticed since PR #3393.  Any PR touching enough shared
files pulls them into the changed-file set and gets blocked by errors it did
not introduce (this is what stalled PR #4507).

Reproduced with the exact CI command from the "Type Check (Mypy)" step of
.github/workflows/ci-standard.yml (mypy==1.13.0, numpy<2.4.0, MYPYPATH
src + src/python/src, --ignore-missing-imports --follow-imports=skip) over
PR #4507's 48-file changed set: 18 errors in 10 files on origin/main,
0 after this change.

Most of the Any leaks come from --follow-imports=skip erasing sibling
modules to Any, so the fix is to restate the real types at those boundaries
rather than to suppress anything:

- calc_backend/routers/thermal_profile.py: bind the pydantic request fields
  to annotated float/str locals (4x no-any-return).
- pressure_drop_calculator/engine/_friction_factors.py,
  _flow_calculations.py: annotate the accumulators fed by Final[float]
  constants (3x no-any-return).
- gui_launcher/launcher.py: annotate the launch_web_app exit code, which is
  declared int at its definition (no-any-return).
- sidekick/process_calculators/optimization.py: declare the evaluate_output
  unpack targets, which the helper already types as
  tuple[float, dict, dict] (2x no-any-return).
- model_generation/converters/simscape/mdl_parser.py: annotate the SLX
  handle as IO[bytes] — it is the zipfile.ZipFile.open() result and the sole
  caller passes exactly that (no-untyped-def).
- psa_package/psa_webapp.py: float() each np.sum before the subtraction,
  matching the three lines above it.  np.sum resolves to Never under these
  stubs, which is what made the "-" operand unsupported (operator).
- plot_engine/trendline.py: float() the np.sum results at the numpy boundary
  (2x var-annotated).
- plot_engine/matplotlib_renderer.py: return from each branch of
  _ensure_fig_ax so every path is provably tuple[Figure, Axes] (return-value).
- model_generation/api/rest_api_routes.py: add the `if content is None`
  guard to the five _extract_text_content call sites that lacked it (3x
  arg-type).  This is the pattern the sixth call site (add_to_library)
  already uses.  `if error: return error` does not narrow `content`, so a
  helper that returned (None, None, ...) would have reached
  MJCFConverter.mjcf_to_urdf / urdf_to_mjcf / URDFParser.parse /
  URDFTextEditor.load_string / SimscapeToURDFConverter.convert_string with
  None and raised instead of returning a 400.

No blanket ignores, no mypy config changes, no baseline file.

Verified: ruff check and ruff format clean on all 10 files; 1072 + 252 +
1532 tests pass across calc_backend, plot_engine, model_generation,
pressure_drop_calculator, sidekick and gui_launcher.  tests/architecture and
tests/ops show the same 8 pre-existing failures before and after.

+++ src/shared/python/plot_engine/trendline.py
@@ -79,9 +79,9 @@ def _r_squared(y: np.ndarray, y_pred: np.ndarray) -> float:
     """Compute R-squared (coefficient of determination)."""
     if y is None:
         raise ValueError("y must be provided")
-    ss_res = np.sum((y - y_pred) ** 2)
-    ss_tot = np.sum((y - np.mean(y)) ** 2)
-    return float(1.0 - (ss_res / ss_tot)) if ss_tot > 0 else 0.0
+    ss_res = float(np.sum((y - y_pred) ** 2))
+    ss_tot = float(np.sum((y - np.mean(y)) ** 2))
+    return 1.0 - (ss_res / ss_tot) if ss_tot > 0 else 0.0
 
 
 def _linear(x: np.ndarray, y: np.ndarray, x_pred: np.ndarray) -> TrendlineResult:
